# swaggerv2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ecommerce RESTful API",
    "version": "2.0.0",
    "description": "A backend RESTful API for an Ecommerce Store",
    "contact": {
      "name": "Custom Haven",
      "url": "https://github.com/CustomHaven"
    },
    "license": {
      "name": "MIT",
      "url": "https://choosealicense.com/licenses/mit/"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4001/",
      "variables": {}
    },
    {
      "url": "https://rest-api-no-orm.herokuapp.com/",
      "variables": {}
    }
  ],
  "paths": {
...
    "/v2/carts": {
      "summary": "Cart Opened",
      "description": "This is the main cart which is linked to all items inside cart-list",
      "get": {
        "summary": "Show all the carts",
        "description": "All the carts only an admin can view this",
        "operationId": "show_cart",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "[{ \"id\": 1, \"user_id\": \"user 1\", \"abandonded\": false, \"created\": \"01-01-2022\", \"modified\": \"04-03-2022\" }, { \"id\": 2, \"user_id\": \"user 2\", \"abandonded\": true, \"created\": \"29-08-2021\", \"modified\": \"11-05-2022\" }, { \"id\": 3, \"user_id\": \"user 3\", \"abandonded\": false, \"created\": \"04-11-2021\", \"modified\": \"12-03-2022\" }, { \"id\": 4, \"user_id\": \"user 2\", \"abandonded\": false, \"created\": \"06-11-2021\", \"modified\": \"28-03-2022\" }, { \"id\": 5, \"user_id\": \"user 3\", \"abandonded\": true, \"created\": \"22-11-2021\", \"modified\": \"28-03-2022\" }]"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"No orders\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
...
    "/v2/carts/{id}": {
      "summary": "Get a cart",
      "description": "Shows the cart to the customer",
      "get": {
        "summary": "Display a cart",
        "description": "Shows the customer their cart",
        "operationId": "display_cart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the cart",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": \"user 3\", \"user_id\": 3, \"abandonded\": false, \"created\": \"04-11-2021\",\n\"modified\": \"12-03-2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Resources not found\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart"
        ]
      },
      "put": {
        "summary": "Update a cart",
        "description": "Update the customer's cart",
        "operationId": "update_cart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the cart",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": \"user 3\", \"user_id\": 3, \"abandonded\": true, \"created\": \"04-11-2021\",\n\"modified\": \"21-08-2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Could not update\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "summary": "Delete a cart",
        "description": "Deletes a cart when it is empty",
        "operationId": "delete_cart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the cart",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"message\": \"Cart deleted\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Could not delete the cart\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/v2/cart-list": {
      "summary": "The list of items in a cart",
      "description": "These paths is all about counting the individual items in the cart their quantity and total prices",
      "get": {
        "summary": "Every item in every cart",
        "description": "This is all search route only for admins showing every item that has been placed in every cart",
        "operationId": "every_cart_item_in_every_cart",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "[{ \"id\": 1, \"product_id\": 1, \"cart_id\": 1, \"quantity\": 6, \"created\": \"26-08-2021\", \"modified\": \"02-04-2022\" }, { \"id\": 2, \"product_id\": 2, \"cart_id\": 1, \"quantity\": 6, \"created\": \"08-09-2021\", \"modified\": \"30-05-2022\" }, { \"id\": 3, \"product_id\": 3, \"cart_id\": 1, \"quantity\": 6, \"created\": \"30-10-2021\", \"modified\": \"03-06-2022\" }, { \"id\": 4, \"product_id\": 4, \"cart_id\": 3, \"quantity\": 4, \"created\": \"12-09-2021\", \"modified\": \"18-04-2022\" }, { \"id\": 5, \"product_id\": 5, \"cart_id\": 3, \"quantity\": 6, \"created\": \"12-11-2021\", \"modified\": \"20-03-2022\" }, { \"id\": 6, \"product_id\": 6, \"cart_id\": 3, \"quantity\": 3, \"created\": \"01-09-2021\", \"modified\": \"28-02-2022\" }, { \"id\": 7, \"product_id\": 7, \"cart_id\": 4, \"quantity\": 4, \"created\": \"31-12-2021\", \"modified\": \"15-05-2022\" }, { \"id\": 8, \"product_id\": 8, \"cart_id\": 4, \"quantity\": 3, \"created\": \"17-12-2021\", \"modified\": \"12-03-2022\" }, { \"id\": 9, \"product_id\": 9, \"cart_id\": 4, \"quantity\": 5, \"created\": \"28-12-2021\", \"modified\": \"09-03-2022\" }, { \"id\": 10, \"product_id\": 10, \"cart_id\": 4, \"quantity\": 1, \"created\": \"09-01-2022\", \"modified\": \"09-03-2022\" }]"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"No items at all\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart List"
        ]
      }
    },
    "/v2/cart-list/cart/{cart_id}": {
      "summary": "The cart with all it's items",
      "description": "Current cart will display all the items that is being added to it",
      "get": {
        "summary": "Cart showing all its items",
        "description": "Route showing all the items that were added to cart",
        "operationId": "cart_with_all_its_items",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "description": "The id of the cart",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\"cart\":{ \"id\": 4, \"user_id\": \"user 2\",\n\"abandonded\": false, \"created\": \"06-11-2021\", \"modified\":\n\"28-03-2022\" }, \"cart_list\":[{ \"id\": 7,\n\"product_id\": 7, \"cart_id\": 4, \"quantity\": 4, \"created\":\n\"31-12-2021\", \"modified\": \"15-05-2022\" }, { \"id\": 8,\n\"product_id\": 8, \"cart_id\": 4, \"quantity\": 3, \"created\":\n\"17-12-2021\", \"modified\": \"12-03-2022\" }, { \"id\": 9,\n\"product_id\": 9, \"cart_id\": 4, \"quantity\": 5, \"created\":\n\"28-12-2021\", \"modified\": \"09-03-2022\" }, { \"id\": 10,\n\"product_id\": 10, \"cart_id\": 4, \"quantity\": 1, \"created\":\n\"09-01-2022\", \"modified\": \"09-03-2022\" }\n]}"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"No items in cart\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart List"
        ]
      },
      "post": {
        "summary": "Add item to cart",
        "description": "Add an item to cart based on the cart id",
        "operationId": "add_item_to_cart",
        "parameters": [
          {
            "name": "cart_id",
            "in": "path",
            "description": "The id of the cart",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart-List"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": 11, \"product_id\": 3, \"cart_id\": 4, \"quantity\": 3, \"created\": \"04-02-2022\", \"modified\": \"07-04-2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Conflict\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart List"
        ]
      }
    },
    "/v2/cart-list/{id}": {
      "summary": "Route for the specific item in cart",
      "description": "Modify the specific item that is inside cart",
      "get": {
        "summary": "Get an item in the cart",
        "description": "Get a specific item in the cart",
        "operationId": "get_a_cart_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the cart item",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": 10, \"product_id\": 10, \"cart_id\": 4, \"quantity\": 3, \"created\": \"09-01-2022\", \"modified\": \"09-03-2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Conflict\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart List"
        ]
      },
      "put": {
        "summary": "Update an item in the cart",
        "description": "Update a specific item in the cart",
        "operationId": "update_cart_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the cart item",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart-List"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": 10, \"product_id\": 10, \"cart_id\": 4, \"quantity\": 3, \"created\": \"09-01-2022\", \"modified\": \"09-03-2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Conflict\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart List"
        ]
      },
      "delete": {
        "summary": "Remove an item from cart",
        "description": "Deletes the item from the cart once quantity reaches 0 done in the frontend",
        "operationId": "delete_cart_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the cart item",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"message\": \"Cart item deleted\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Could not delete the item from the cart\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Cart List"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user_id": {
            "type": "string"
          },
          "abandonded": {
            "type": "boolean"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Cart"
        }
      },
      "Cart-List": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "product_id": {
            "type": "string"
          },
          "cart_id": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Cart-List"
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect_sid"
      }
    }
  }
}